new file:   .vscode/launch.json 	modified:   algo_new/Double_UCB.py 	modified:   algo_new/Double_UCB_BL.py 	new file:   algo_new/Unkown_UCB_BL.py 	modified:   aux_func.py 	modified:   param.py 	modified:   test_two_ball.py

+++ algo_new/Double_UCB.py
@@ -7,13 +7,13 @@ from numba import jit
 @jit(nopython=True)
 def Double_UCB(T,reward_mat,arm_type,neighbor_init,change_arms_list,seed):
     np.random.seed(seed)
-    N=np.zeros(T,dtype='int32')
+    N=np.zeros(T,dtype='int64')
     X_hat=np.zeros(T)
     c=np.zeros(T)
     c_add_optimal=np.zeros(T)
     c_add=np.zeros(T)
     c_sub=np.zeros(T)
-    N_optimal=np.zeros(T,dtype='int32')
+    N_optimal=np.zeros(T,dtype='int64')
     
     arm_ind=np.array([0])
     X_hat_optimal=np.zeros(T)
@@ -89,13 +89,13 @@ def Double_UCB(T,reward_mat,arm_type,neighbor_init,change_arms_list,seed):
 @jit(nopython=True)
 def C_UCB(T,reward_mat,arm_type,neighbor_init,change_arms_list,seed):
     np.random.seed(seed)
-    N=np.zeros(T,dtype='int32')
+    N=np.zeros(T,dtype='int64')
     X_hat=np.zeros(T)
     c=np.zeros(T)
     c_add_optimal=np.zeros(T)
     c_add=np.zeros(T)
     c_sub=np.zeros(T)
-    N_optimal=np.zeros(T,dtype='int32')
+    N_optimal=np.zeros(T,dtype='int64')
     
     arm_ind=np.array([0])
     X_hat_optimal=np.zeros(T)
